requirements.txt
astroid==2.2.5
autopep8==1.4.3
isort==4.3.16
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
